Adding note (#24)
When reporting to clients, they tend not to care (nor wish) to see the code. This means it's common to set knitr options such that `echo=FALSE`. With this option set, arbitrary marginfigure chunk content is not shown and this fact is not otherwise documented. However, it's easily fixed by add echo=TRUE to the marginfigure chunk option. I do not know if this effects the LaTeX version, as I don't use latex.